requirements.txt
langchain
langgraph
langchain-openai
langchain-community
chromadb
pydantic
python-dotenv